微信小程序用户端 1、完成商品详情页的展示相关接口 2、完成展示商品评论的接口 3、完成购物车模块搭建 4、完成查询当前登录用户的购物车内商品数量接口，展示到购物车图表右上角 5、完成购物车列表展示接口（部分完成，RPC查询未完成） 6、解决了一些其他bug

+++ power-business/search-service/src/main/java/com/chuanqihou/powershop/controller/SearchProdController.java
@@ -1,7 +1,6 @@
 package com.chuanqihou.powershop.controller;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.chuanqihou.powershop.domain.ProdTag;
 import com.chuanqihou.powershop.model.ProdEs;
 import com.chuanqihou.powershop.model.Result;
 import com.chuanqihou.powershop.service.SearchProdService;
@@ -23,7 +22,7 @@ public class SearchProdController {
     private SearchProdService searchProdService;
 
     @GetMapping("/prodListByTagId")
-    public Result<Page<ProdEs>> getProdTageById(Long tagId,Long size) {
+    public Result<Page<ProdEs>> getProdTageById(Long tagId,Integer size) {
         Page<ProdEs> prodEsPage = searchProdService.searchProdEsByTagIdPage(tagId,size);
         return Result.success(prodEsPage);
     }
